almost done

+++ app/src/main/java/app/budgetku/domain/usecase/wallet/DeleteWalletUseCase.java
@@ -12,7 +12,11 @@ public class DeleteWalletUseCase extends DeleteUseCase<Wallet> {
     }
 
     @Override
-    public void execute(Wallet data) {
-        repository.deleteWallet(data);
+    public void execute(Wallet data) throws Exception {
+        if (repository.getWallets().size() < 2) {
+            throw new Exception("The last available wallet cannot be deleted");
+        } else {
+            repository.deleteWallet(data);
+        }
     }
 }
